Bulk-tab update
