+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Router } from "./routes";
 import { AuthenticationWrapper } from "./components/Account";
-import { MenuKey } from "./components/MenuKey";
+import { MenuKey } from "./components";
 
 const App = () => (
   <MenuKey>
