+++ src/components/index.js
@@ -1,7 +1,8 @@
 import { Login, Register } from "./Account";
 import { Dashboard } from "./Dashboard";
 import { ErrorPage } from "./ErrorPage";
+import { MenuKey } from "./MenuKey";
 import { Profile } from "./Profile";
 
-export { Dashboard, ErrorPage, Login, Profile, Register };
+export { Dashboard, ErrorPage, Login, MenuKey, Profile, Register };
