+++ src/modules/job/Create/Bulk/index.js
@@ -1,0 +1,1 @@
+export { default as BulkUpload } from "./Bulk";
