+++ src/modules/job/Create/Single/index.js
@@ -1,0 +1,1 @@
+export { default as SingleUpload } from "./Single";
